refactor(cli): split InteractiveConsoleApp.Shell.cs into 4 partials
Shell.cs was 2178 lines and growing — well past the codebase ceiling.
Split into:

- Shell.cs (1148) — orchestration: RunAsync, CreateCommandCenter, status
  bars, modals, palette, theme, page-shortcut helpers, async callbacks
- Home.cs (80) — BuildHomePage only
- Catalog.cs (468) — Skills, Collections, Bundles, Packages, Agents
- Workspace.cs (541) — Installed, Project, Analysis, Settings,
  RemoveAll/UpdateAll confirmations, About

Zero behavior change. All partials of the same internal sealed class.
Build clean, 0 warnings across all 4 projects. 613/613 tests pass.
